requirements.txt
asgiref==3.2.5
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
covid==2.2.1
covid-data-api==1.2
COVID19Py==0.3.0
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.4
django-import-export==2.0.2
django-social-share==1.4.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.9
jdcal==1.4.1
kiwisolver==1.2.0
lxml==4.5.0
MarkupPy==1.14
matplotlib==3.2.1
numpy==1.18.2
odfpy==1.4.1
openpyxl==3.0.3
panda==0.3.1
pandas==1.0.3
pydantic==1.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
stripe==2.45.0
tablib==1.1.0
urllib3==1.25.8
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
